# Conflicts:
#	src/main/java/com/teamChallenge/controller/AdsController.java
#	src/main/java/com/teamChallenge/controller/AuthController.java
#	src/main/java/com/teamChallenge/controller/BannerController.java
#	src/main/java/com/teamChallenge/controller/CartController.java
#	src/main/java/com/teamChallenge/controller/CategoryController.java
#	src/main/java/com/teamChallenge/controller/FigureController.java
#	src/main/java/com/teamChallenge/controller/OrderController.java
#	src/main/java/com/teamChallenge/controller/ReviewController.java
#	src/main/java/com/teamChallenge/controller/SubCategoryController.java
#	src/main/java/com/teamChallenge/controller/UserController.java
#	src/main/java/com/teamChallenge/controller/user/WishListController.java

+++ src/main/java/com/teamChallenge/controller/AdsController.java
@@ -1,10 +1,15 @@
 package com.teamChallenge.controller;
 
 import com.teamChallenge.dto.request.AdsRequestDto;
+import com.teamChallenge.dto.request.figure.FigureRequestDto;
 import com.teamChallenge.dto.response.AdsResponseDto;
-import com.teamChallenge.entity.advertisement.AdvertisementService;
+import com.teamChallenge.dto.response.FigureResponseDto;
+import com.teamChallenge.entity.advertisement.AdvertisementServiceImpl;
 import com.teamChallenge.exception.CustomErrorResponse;
 import com.teamChallenge.exception.LogEnum;
+import com.teamChallenge.exception.exceptions.generalExceptions.CustomNotFoundException;
+import com.teamChallenge.exception.exceptions.generalExceptions.UnauthorizedAccessException;
+import com.teamChallenge.security.AuthUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -13,34 +18,44 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 
 @Slf4j
+@Validated
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/ads")
 public class AdsController {
 
     private final AdvertisementService adsService;
+    private static final String URI_WITH_ID = "/{id}";
+    private static final String SEC_REC = "BearerAuth";
 
-    public static final String URI_ADS_WITH_ID = "/{id}";
-
-    @PostMapping
+    @PostMapping()
+    @SecurityRequirement(name = SEC_REC)
     @Operation(summary = "Add new Ads")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "201", description = "Added new Ads",
-                    content = { @Content(mediaType = "application/json",
+                    content = { @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                             schema = @Schema(implementation = AdsResponseDto.class))}),
             @ApiResponse(responseCode = "400", description = "Validation errors",
-                    content = { @Content(mediaType = "application/json",
+                    content = { @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                             schema = @Schema(implementation = CustomErrorResponse.class))})
     })
-    @SecurityRequirement(name = "BearerAuth")
-    public AdsResponseDto addAds(@Valid @RequestBody AdsRequestDto request) {
+    public AdsResponseDto create(@Valid @NotNull @RequestBody AdsRequestDto request, Principal principal) throws UnauthorizedAccessException {
+        authUtil.validateAdminRole(principal);
+
         AdsResponseDto ads = adsService.create(request);
         log.info("{}: Figure (id: {}) has been added", LogEnum.SERVICE, ads.id());
         return ads;
@@ -50,42 +65,64 @@ public class AdsController {
     @Operation(summary = "Get all ads")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "List of ads",
-                    content = { @Content(mediaType = "application/json",
+                    content = { @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                             array = @ArraySchema(schema = @Schema(implementation = AdsResponseDto.class)))}
             )
     })
     public List<AdsResponseDto> adsList() {
         List<AdsResponseDto> adsDtoList = adsService.getAll();
+
         log.info("{}: Ads list have been retrieved", LogEnum.CONTROLLER);
         return adsDtoList;
     }
 
-    @GetMapping(URI_ADS_WITH_ID)
+    @GetMapping(URI_WITH_ID)
     @Operation(summary = "Get ads by ID")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Got ads",
-                    content = { @Content(mediaType = "application/json",
+            @ApiResponse(responseCode = "200", description = "Ads received",
+                    content = { @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                             schema = @Schema(implementation = AdsResponseDto.class)) }),
             @ApiResponse(responseCode = "404", description = "Ads not found",
-                    content = { @Content(mediaType = "application/json",
+                    content = { @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                             schema = @Schema(implementation = CustomErrorResponse.class)) })
 
     })
-    public AdsResponseDto getAdsById(@PathVariable String id) {
+    public AdsResponseDto getById(@NotBlank @NotNull @PathVariable String id) {
         AdsResponseDto ads = adsService.getById(id);
         log.info("{}: Ads (id: {}) has been retrieved", LogEnum.SERVICE, ads.id());
         return ads;
     }
 
-    @DeleteMapping(URI_ADS_WITH_ID)
+    @PutMapping(URI_WITH_ID)
+    @SecurityRequirement(name = SEC_REC)
+    @Operation(summary = "Update ads")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Ads updated"),
+            @ApiResponse(responseCode = "404", description = "Ads not found",
+                    content = { @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = CustomErrorResponse.class))})
+    })
+    public AdsResponseDto update(@PathVariable String id, @NotNull @RequestBody AdsRequestDto adsDto, Principal principal) throws CustomNotFoundException, UnauthorizedAccessException{
+        authUtil.validateAdminRole(principal);
+
+        AdsResponseDto updated = adsService.update(id, adsDto);
+        log.info("{}: Ads (id: {}) has been updated", LogEnum.CONTROLLER, id);
+
+        return updated;
+    }
+
+    @DeleteMapping(URI_WITH_ID)
+    @SecurityRequirement(name = SEC_REC)
+    @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "Delete ads")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Ads deleted"),
             @ApiResponse(responseCode = "404", description = "Ads not found",
-                    content = { @Content(mediaType = "application/json",
+                    content = { @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
                             schema = @Schema(implementation = CustomErrorResponse.class)) }) })
-    @SecurityRequirement(name = "BearerAuth")
-    public void deleteAdsById(@PathVariable String id) {
+    public void delete(@PathVariable String id, Principal principal) throws UnauthorizedAccessException {
+        authUtil.validateAdminRole(principal);
+
         adsService.delete(id);
         log.info("{}: Ads (id: {}) has been deleted", LogEnum.CONTROLLER, id);
     }
